+++ teensy40_brushed_v2/lib/Arm/driver_motor.h
@@ -198,7 +198,7 @@ public:
 	 *
 	 * @param		speed in range -1 to 1
 	 */
-	void move_manual(uint32_t speed);
+	void move_manual(float speed);
 
 	float get_current_angle_es();
 	float get_current_angle_ps();
